Add file uploads.
This uses fieldname@filename to create a multipart form upload.  For the
moment, the content-type of uploaded files is set to
application/octet-stream.

The code to handle this and the regular form handling code is a bit
ugly.  Not sure exactly how to refactor, but this system is going to
need a testing framework soon to avoid regressions and unintentional
changes in behaviour.