Simplify adding key-value pairs to request
We used to maybe return a nil pointer, which required extra checks when
we were looking at the result of the function.  Now, we *always* return
a valid pointer, and all the maps are filed in (although sometimes
empty).  This removes the need for a number of check we were doing
before we could use the data provided.